refactor controller utilities
